requirements.txt
PyPDF2
langchain==0.0.263
python-dotenv==1.0.0
streamlit==1.22.0
pandas 
numpy
matplotlib
notebook
pinecone-client==2.2.4
fastapi
pydantic
uvicorn
google-generativeai
datasets
transformers
sentence-transformers
pdfplumber
spacy